requirements.txt
gpt4all==2.0.2
python-dotenv==1.0.1
redis==5.0.1
Requests==2.31.0
